Give dbname explicitly to psql
## Problem

Rspec fails when 'default' database for psql does not exist.

## Cause

If you do not specify dbname for psql, psql trys to connect to a database which has the same name with current user. If there does not exist such a database, psql fails to connect and exits.
Due to this psql's behaviour, `restore_database_postgresql` and `restore_tables_postgresql` fail in some environments.

## Remedy

* Specify dbname 'ridgepole_test' explicitly in `restore_database_postgresql` and `restore_tables_postgresql`
* Check exit status of psql and raise a RuntimeError on failure.