Add `logger` gem as an explicit dependency
`logger` will be a bundled gem from Ruby 3.5.
https://bugs.ruby-lang.org/issues/20309

So if you use `logger` as the standard library, Bundler shows the
following warning.

```
warning: logger was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
```

I'm not sure why yet, but it seems that the above message has showsn
since Ruby 3.3.5. This message has broken specs.
https://github.com/ridgepole/ridgepole/actions/runs/10765746072/job/29850451424?pr=501

Regardless of the cause, we need to add `logger` to a dependency to
keep using `logger`.  This PR adds `logger` gem as a runtime dependency.